+++ R/BumpDisconnectCells.R
@@ -29,6 +29,8 @@
 #'
 #' @keywords utilities
 #'
+#' @importClassesFrom raster BasicRaster
+#'
 #' @export
 #'
 #' @examples